+++ _examples/configuration/functional/main.go
@@ -15,9 +15,9 @@ func main() {
 	// Prefix: "With", code editors will help you navigate through all
 	// configuration options without even a glitch to the documentation.
 
-	app.Listen(":8080", iris.WithoutStartupLog, iris.WithCharset("UTF-8"))
+	app.Listen(":8080", iris.WithoutStartupLog, iris.WithCharset("utf-8"))
 
 	// or before run:
-	// app.Configure(iris.WithoutStartupLog, iris.WithCharset("UTF-8"))
+	// app.Configure(iris.WithoutStartupLog, iris.WithCharset("utf-8"))
 	// app.Listen(":8080")
 }