+++ _examples/view/template_html_0/main.go
@@ -24,7 +24,7 @@ func main() {
 	app.Get("/", hi)
 
 	// http://localhost:8080
-	app.Listen(":8080", iris.WithCharset("UTF-8")) // defaults to that but you can change it.
+	app.Listen(":8080", iris.WithCharset("utf-8")) // defaults to that but you can change it.
 }
 
 func hi(ctx iris.Context) {